+++ CodeMirror6/Commands/CMCommandDispatcher.cs
@@ -23,4 +23,13 @@ public class CMCommandDispatcher
     /// <param name="value"></param>
     /// <returns></returns>
     public Task Dispatch<TValue>(CodeMirrorCommandOneParameter command, TValue value) => cmJsInterop.ModuleInvokeVoidAsync("dispatchCommand", command, value);
+
+    /// <summary>
+    /// Invoke a built-in CodeMirror command with two parameters
+    /// </summary>
+    /// <param name="command"></param>
+    /// <param name="value1"></param>
+    /// <param name="value2"></param>
+    /// <returns></returns>
+    public Task Dispatch<TValue1, TValue2>(CodeMirrorCommandTwoParameters command, TValue1 value1, TValue2 value2) => cmJsInterop.ModuleInvokeVoidAsync("dispatchCommand", command, value1, value2);
 }