+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 0.3.0 - 2024-02-06
+
+### ✨ Introduce new features
+
+- Add commands to insert markdown table and horizontal rule
+- Implement csv and tsv with padded column widths (Does not support multiline cells)
+- Add Tab and Shift-Tab keymaps to csv mode, to navigate columns
+
+### 🎨 Improve structure / format of the code
+
+- Improve eq() of emojiWidget
+
 ## 0.2.2 - 2024-02-01
 
 ### ⚡️ Improve performance